add first part of q1 hw3

+++ Lecture2/delete.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.19;
 
 contract deleteItems {
 
-    uint256[] public items = [0,1,2,3,4,5,6,7,8,9,10,11,12]; 
+    uint[] public items = [0,1,2,3,4,5,6,7,8,9,10,11,12]; 
  
     function randomNumber() public view returns(uint) {
         return uint(keccak256(abi.encodePacked(block.prevrandao, block.timestamp, items.length)));
